+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Test now ensures html_root_url is present
 * Fixed compatibility with Miri
 * Adjusted contribution guidelines
+* AsRef and Deref are now implemented via those respective traits on owned values
 
 ## 0.0.5